Move ToArray to the algorithm folder
