+++ Plugins/GridBased2D/Source/GridBased2D/Private/Components/GridBasedMovementComponent.cpp
@@ -10,7 +10,7 @@
 #include "Map/TileMapGridBasedMap.h"
 #include "MathUtilities.h"
 #include "Ranges/Views/ContainerView.h"
-#include "Ranges/Views/ToArray.h"
+#include "Ranges/Algorithm/ToArray.h"
 #include <range/v3/view/filter.hpp>
 #include <range/v3/view/transform.hpp>
