+++ Plugins/PokemonStarterKit/Source/PokemonBattle/Private/Battle/Actions/BattleActionUseItem.cpp
@@ -12,7 +12,7 @@
 #include "DataTypes/OptionalUtilities.h"
 #include "PokemonBattleSettings.h"
 #include "Ranges/Views/ContainerView.h"
-#include "Ranges/Views/ToArray.h"
+#include "Ranges/Algorithm/ToArray.h"
 #include <range/v3/view/filter.hpp>
 #include <range/v3/view/transform.hpp>
