+++ Plugins/PokemonStarterKit/Source/PokemonData/Private/Bag/Item.cpp
@@ -5,7 +5,7 @@
 #include "range/v3/view/filter.hpp"
 #include "range/v3/view/transform.hpp"
 #include "Ranges/Views/ContainerView.h"
-#include "Ranges/Views/ToArray.h"
+#include "Ranges/Algorithm/ToArray.h"
 
 FItem::FItem() = default;
