+++ Plugins/PokemonStarterKit/Source/PokemonData/Private/Battle/Status.cpp
@@ -2,7 +2,7 @@
 #include "Battle/Status.h"
 #include "DataManager.h"
 #include "Ranges/Views/ContainerView.h"
-#include "Ranges/Views/ToArray.h"
+#include "Ranges/Algorithm/ToArray.h"
 #include <range/v3/view/transform.hpp>
 
 TArray<FName> UStatusEffectHelper::GetStatusEffectNames() {
