+++ Plugins/PokemonStarterKit/Source/PokemonData/Private/Species/Stat.cpp
@@ -3,7 +3,7 @@
 #include "Algo/RemoveIf.h"
 #include "DataManager.h"
 #include "Ranges/Views/ContainerView.h"
-#include "Ranges/Views/ToArray.h"
+#include "Ranges/Algorithm/ToArray.h"
 #include <range/v3/view/filter.hpp>
 #include <range/v3/view/transform.hpp>
